refactor: Rename `lib/plugins/aws/lib/checkIfEcrRepositoryExists.js`
