+++ lib/plugins/aws/remove/index.js
@@ -6,7 +6,7 @@ const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const emptyS3Bucket = require('./lib/bucket');
 const removeStack = require('./lib/stack');
 const removeEcrRepository = require('./lib/ecr');
-const checkIfEcrRepositoryExists = require('../lib/checkIfEcrRepositoryExists');
+const checkIfEcrRepositoryExists = require('../lib/check-if-ecr-repository-exists');
 const { log, style, progress } = require('@serverless/utils/log');
 
 const mainProgress = progress.get('main');
